% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestNormalize-package.R
\docType{package}
\name{bestNormalize-package}
\alias{bestNormalize-package}
\alias{_PACKAGE}
\title{bestNormalize: Flexibly calculate the best normalizing transformation for a
vector}
\description{
The \code{bestNormalize} package provides several normalizing transformations, and introduces a
new transformation based off of the order statistics, \code{orderNorm}.
Perhaps the most useful function is \code{bestNormalize}, which attempts all
of these transformations and picks the best one based off of a goodness of
fit statistic.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/petersonR/bestNormalize}
}

}
\author{
\strong{Maintainer}: Ryan Andrew Peterson \email{ryan.a.peterson@cuanschutz.edu}

}
