% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestNormalize-package.R, R/bestNormalize.R
\docType{package}
\name{bestNormalize}
\alias{bestNormalize}
\alias{bestNormalize-package}
\alias{bestNormalize}
\alias{predict.bestNormalize}
\alias{predict.bestNormalize}
\alias{print.bestNormalize}
\title{bestNormalize: Flexibly calculate the best normalizing transformation for a vector}
\usage{
bestNormalize(x, allow_orderNorm = TRUE, ...)

\method{predict}{bestNormalize}(BNobject, newdata = NULL, inverse = FALSE)

\method{print}{bestNormalize}(BNobject)
}
\arguments{
\item{x}{A vector to normalize}

\item{allow_orderNorm}{set to FALSE if orderNorm should not be applied}
}
\value{
A list of class \code{bestNormalize} with elements 

\item{x.t}{transformed original data} 
\item{x}{original data} 
\item{norm_stats}{Pearson's chi-squared normality test statistics}
\item{chosen_transform}{info about the tranformation (of appropriate class)}

The \code{predict} function returns the numeric value of the transformation
performed on new data, and allows for the inverse transformation as well.
}
\description{
The foo package provides several normalizing tranformations, and introduces a new transformation based
off of the order statistics, \code{orderNorm}. Perhaps the most useful function is 
\code{bestNormalize}, which attempts all of these transformations and picks the best one 
based off of a goodness of fit statistic.
#'

Performs a suite of normalizing transformations, and selects the
  best one on the basis of the Shapiro-Wilks test for normality (ie, the one 
  with the highest p-value, ie the one that shows the least amount of
  evidence against normality).
}
\details{
\code{bestNormalize} estimates the optimal normalizing 
  transformation. This transformation can be performed on new data, and 
  inverted, via the \code{predict} function.
}
\examples{

x <- rgamma(100, 1, 1)

BN_obj <- bestNormalize(x)
BN_obj
p <- predict(BN_obj)
x2 <- predict(BN_obj, newdata = p, inverse = TRUE)

all.equal(x2, x)

}
\seealso{
\code{\link[bestNormalize]{boxcox}},
 \code{\link{lambert}}, 
 \code{\link{orderNorm}},
 \code{\link{yeojohnson}}
}
